make Union::try_new pub (#18125)
## Which issue does this PR close?

- Closes #18126.

## Rationale for this change

It's a useful constructor for users manipulating logical plans where
they know the schemas will match exactly. We already expose other
constructors for Union and constructors for logical plans.

## What changes are included in this PR?

Makes `Union::try_new` a public function.

## Are these changes tested?

Seems unnecessary.

## Are there any user-facing changes?

The function is now public. Not a breaking change, but going forward
changes to it would breaking changes to users of the logical plan API.