fix: window unparsing (#17367)
## Which issue does this PR close?

- Closes #17360.

## Rationale for this change

in LogicalPlan::Filter unparsing, 
if there's a window expr, it should be converted to quailify.

postgres must has an alias for derived table. otherwise it will
complain:
```
ERROR: subquery in FROM must have an alias.
```
fixed this issue at the same time.

## What changes are included in this PR?

If window expr is found, convert filter to quailify.

## Are these changes tested?

UT

## Are there any user-facing changes?

No

---------

Co-authored-by: Jeffrey Vo <jeffrey.vo.australia@gmail.com>