fix(substrait): schema errors for Aggregates with no groupings (#17909)
## Which issue does this PR close?
Closes https://github.com/apache/datafusion/issues/16590

## Rationale for this change
When consuming Substrait plans containing aggregates with no groupings,
we would see the following error
```
Error: Substrait("Named schema must contain names for all fields")
```

The Substrait plan had one _less_ field than DataFusion expected because
DataFusion was adding an extra "__grouping_id" to the output of the
Aggregate node. This happens when the

https://github.com/apache/datafusion/blob/daeb6597a0c7344735460bb2dce13879fd89d7bd/datafusion/expr/src/logical_plan/plan.rs#L3418
condition is true.

A natural followup question to this is "Why are we creating an Aggregate
with a single empty GroupingSet for the group by, instead of just
leaving the group by entirely?".

## What changes are included in this PR?
Instead of setting group_exprs to a vector with a single empty grouping
set, let's just leave group_exprs empty entirely. This means that the
`is_grouping_set` is not triggered, so the Datafusion schema matches the
Substrait schema.

## Are these changes tested?
Yes

I have added direct tests via example Substrait plans

## Are there any user-facing changes?
Substrait plans that were not consumable before are now consumable.

---------

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>