Insta for enforce_distrubution (easy ones) (#18248)
- part of https://github.com/apache/datafusion/issues/15791

All easy cases from https://github.com/apache/datafusion/pull/18185
(that are nicely-ish displayed in git diff).

Note on preserving comments: if it was note about what should happen (or
what will be tested), it's placed on top of the snapshot. If that's
something that comments part of the plan, I put it below the plan